+++ Libplanet.Net.Tests/Consensus/Context/ContextProposerTest.cs
@@ -314,7 +314,8 @@ namespace Libplanet.Net.Tests.Consensus.Context
             var block = blockChain.ProposeBlock(new PrivateKey(), DateTimeOffset.UtcNow);
             blockChain.Append(block, TestUtils.CreateBlockCommit(block));
             Assert.Equal(
-                TestUtils.PrivateKeys[2].PublicKey, TestUtils.ValidatorSet.GetProposer(2, 0));
+                TestUtils.PrivateKeys[2].PublicKey,
+                TestUtils.ValidatorSet.GetProposer(2, 0).PublicKey);
 
             context.Start();
             await exceptionOccurred.WaitAsync();